+++ robotiq_ft_driver/src/robotiq_ft_driver_node_main.ros2.cpp
@@ -1,0 +1,16 @@
+#include <memory>
+
+#include <rclcpp/rclcpp.hpp>
+
+#include "robotiq_ft_driver_node.ros2.hpp"
+
+int main(int argc, char * argv[])
+{
+  auto args = rclcpp::init_and_remove_ros_arguments(argc, argv);
+  using robotiq_ft_driver::RobotiqFTDriverNode;
+  rclcpp::spin(
+      std::make_shared<RobotiqFTDriverNode>(
+          rclcpp::NodeOptions().arguments(args)));
+  rclcpp::shutdown();
+  return 0;
+}
